add html css and bootstrap for the project
This is a sample page project with a responsive website. I used for project HTML CSS and BOOTSTRAP. In this project, there is a valid contact form with an email address. If any visitor would like to connect with you he or she can do it by name and an email.